Refactor Algolia search into it's own JS file.
